permissions

+++ src/app/Http/Controllers/PermissionController.php
@@ -4,9 +4,10 @@ namespace IBoot\Core\App\Http\Controllers;
 
 use App\Http\Controllers\Controller;
 use IBoot\Core\App\Exceptions\ServerErrorException;
-use IBoot\Core\App\Http\Requests\RoleRequest;
+use IBoot\Core\App\Http\Requests\PermissionRequest;
 use IBoot\Core\App\Services\PermissionService;
 use IBoot\Core\App\Services\RoleService;
+use Spatie\Permission\Models\Permission;
 use Illuminate\Contracts\View\View;
 use Illuminate\Http\Request;
 use Illuminate\Http\JsonResponse;
@@ -37,6 +38,7 @@ class PermissionController extends Controller
      */
     public function index(): View|string
     {
+        $this->authorize('viewAny', Permission::class);
         $permissions = $this->permission->getLists();
 
         return view('packages/core::permissions.index', compact('permissions'));
@@ -44,6 +46,7 @@ class PermissionController extends Controller
 
     public function create(): View
     {
+        $this->authorize('create', Permission::class);
         $roles = $this->role->getLists();
 
         return view('packages/core::permissions.form', compact('roles'));
@@ -54,6 +57,7 @@ class PermissionController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize('edit', Permission::class);
         $permission = $this->permission->getById($id)->load('roles');
         $roles = $this->role->getLists();
 
@@ -67,7 +71,7 @@ class PermissionController extends Controller
      * @return JsonResponse
      * @throws ServerErrorException
      */
-    public function update(RoleRequest $request, string $id): JsonResponse
+    public function update(PermissionRequest $request, string $id): JsonResponse
     {
         DB::beginTransaction();
         try {
@@ -91,6 +95,7 @@ class PermissionController extends Controller
     {
         DB::beginTransaction();
         try {
+            $this->authorize('delete', Permission::class);
             $this->permission->deletePermission($id);
             DB::commit();
