+++ src/app/Http/Controllers/RoleController.php
@@ -7,6 +7,7 @@ use IBoot\Core\App\Exceptions\ServerErrorException;
 use IBoot\Core\App\Http\Requests\RoleRequest;
 use IBoot\Core\App\Services\PermissionService;
 use IBoot\Core\App\Services\RoleService;
+use Spatie\Permission\Models\Role;
 use Illuminate\Contracts\View\View;
 use Illuminate\Http\Request;
 use Illuminate\Http\JsonResponse;
@@ -37,6 +38,7 @@ class RoleController extends Controller
      */
     public function index(): View|string
     {
+        $this->authorize('viewAny', Role::class);
         $roles = $this->role->getLists();
 
         return view('packages/core::roles.index', compact('roles'));
@@ -44,6 +46,7 @@ class RoleController extends Controller
 
     public function create(): View
     {
+        $this->authorize('create', Role::class);
         $permissions = $this->permission->getLists();
 
         return view('packages/core::roles.form', compact('permissions'));
@@ -54,6 +57,7 @@ class RoleController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize('edit', Role::class);
         $role = $this->role->getById($id)->load('permissions');
         $permissions = $this->permission->getLists();
 
@@ -91,6 +95,7 @@ class RoleController extends Controller
     {
         DB::beginTransaction();
         try {
+            $this->authorize('delete', Role::class);
             $this->role->deleteRole($id);
             DB::commit();
